Global replace of integral single-precision literals with integer literals.
As part of a large clean-up of the code to align with our recent decision to use (CGFloat)0.5 style literals, the first step is to migrate as many values to integer literals as possible. These are easier to read than casts to CGFloat and will be promoted to double-precision floating point numbers without loss of precision.

PiperOrigin-RevId: 220622292

+++ components/FlexibleHeader/examples/FlexibleHeaderFABExample.m
@@ -18,7 +18,7 @@
 #import "MaterialFlexibleHeader.h"
 #import "supplemental/FlexibleHeaderFABSupplemental.h"
 
-static const CGFloat kFlexibleHeaderMinHeight = 200.f;
+static const CGFloat kFlexibleHeaderMinHeight = 200;
 
 @interface FlexibleHeaderFABExample () <UIScrollViewDelegate>
 
@@ -81,7 +81,7 @@ static const CGFloat kFlexibleHeaderMinHeight = 200.f;
   [self.view addSubview:self.fhvc.view];
   [self.fhvc didMoveToParentViewController:self];
 
-  self.fhvc.headerView.backgroundColor = [UIColor colorWithWhite:0.1f alpha:1.0f];
+  self.fhvc.headerView.backgroundColor = [UIColor colorWithWhite:0.1f alpha:1];
 }
 
 - (void)viewWillAppear:(BOOL)animated {
@@ -92,9 +92,9 @@ static const CGFloat kFlexibleHeaderMinHeight = 200.f;
   [self.navigationController setNavigationBarHidden:YES animated:animated];
 
   self.floatingButton = [[MDCFloatingButton alloc] init];
-  [self.floatingButton setBackgroundColor:[UIColor colorWithRed:11/255.0f
-                                                          green:232/255.0f
-                                                           blue:94/255.0f
+  [self.floatingButton setBackgroundColor:[UIColor colorWithRed:11 / (CGFloat)255
+                                                          green:232 / (CGFloat)255
+                                                           blue:94 / (CGFloat)255
                                                           alpha:1]
                                  forState:UIControlStateNormal];
   [self.floatingButton sizeToFit];
